+++ CHANGELOG.md
@@ -19,6 +19,9 @@ We use the following categories for changes:
 - `-enable-feature=promql-per-step-stats` feature for statistics in PromQL evaluation
 - Add `readinessProbe` in helm chart [#1266]
 
+### Fixed
+- Fix trace querying by status code tag [#1384]
+
 ## [0.11.0] - 2022-05-11
 
 ## Added